requirements.txt
Flask
flask-cors
mysql-connector-python
textblob
nltk
waitress
python-dotenv
groq
